# DatametrixSolution/Core.Domain/Entities/Task.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Core.Domain.Entities
{
    public class Task
    {
        public int Id { get; set; }
        public string Name { get; set; }
    }
}